feat: generate anonymous participation url modal
replaces content in anon recruitment section. modal triggered by
 button in the section. disabled if anon recruitment is disabled
